Sync master with develop (squash merge from develop)

+++ src/groups/services/groups.service.ts
@@ -1,6 +1,5 @@
-import { Injectable, Logger, Inject } from '@nestjs/common';
+import { Injectable, Logger, Inject } from "@nestjs/common";
 import {
-  FindConditions,
   ILike,
   In,
   LessThanOrEqual,
@@ -8,41 +7,43 @@ import {
   Repository,
   Connection,
   TreeRepository,
-} from 'typeorm';
-import Group from '../entities/group.entity';
-import GroupEvent from '../../events/entities/event.entity';
-import SearchDto from '../../shared/dto/search.dto';
-import { GroupSearchDto } from '../dto/group-search.dto';
-import GroupListDto from '../dto/group-list.dto';
-import CreateGroupDto from '../dto/create-group.dto';
-import UpdateGroupDto from '../dto/update-group.dto';
-import { GroupDetailDto } from '../dto/group-detail.dto';
-import GooglePlaceDto from '../../vendor/google-place.dto';
-import { GoogleService } from '../../vendor/google.service';
-import ClientFriendlyException from '../../shared/exceptions/client-friendly.exception';
-import GroupMembership from '../entities/groupMembership.entity';
-import { GroupRole } from '../enums/groupRole';
-import { hasValue } from '../../utils/validation';
-import GroupCategoryReport from '../entities/groupCategoryReport.entity';
-import { endOfMonth, startOfMonth } from 'date-fns';
+} from "typeorm";
+import Group from "../entities/group.entity";
+import GroupEvent from "../../events/entities/event.entity";
+import SearchDto from "../../shared/dto/search.dto";
+import { GroupSearchDto } from "../dto/group-search.dto";
+import GroupListDto from "../dto/group-list.dto";
+import CreateGroupDto from "../dto/create-group.dto";
+import UpdateGroupDto from "../dto/update-group.dto";
+import { GroupDetailDto } from "../dto/group-detail.dto";
+import GooglePlaceDto from "../../vendor/google-place.dto";
+import { GoogleService } from "../../vendor/google.service";
+import ClientFriendlyException from "../../shared/exceptions/client-friendly.exception";
+import GroupMembership from "../entities/groupMembership.entity";
+import { GroupRole } from "../enums/groupRole";
+import { hasValue } from "../../utils/validation";
+import { endOfMonth, startOfMonth } from "date-fns";
+import { GroupPermissionsService } from "./group-permissions.service";
+import GroupCategory from "../entities/groupCategory.entity";
 
 @Injectable()
 export class GroupsService {
   private readonly repository: Repository<Group>;
   private readonly treeRepository: TreeRepository<Group>;
   private readonly membershipRepository: Repository<GroupMembership>;
-  private readonly groupReportRepository: Repository<GroupCategoryReport>;
   private readonly eventRepository: Repository<GroupEvent>;
+  private readonly groupCategoryRepository: Repository<GroupCategory>;
 
   constructor(
-    @Inject('CONNECTION') connection: Connection,
+    @Inject("CONNECTION") connection: Connection,
+    private groupsPermissionsService: GroupPermissionsService,
     private googleService: GoogleService,
   ) {
     this.repository = connection.getRepository(Group);
     this.treeRepository = connection.getTreeRepository(Group);
     this.membershipRepository = connection.getRepository(GroupMembership);
-    this.groupReportRepository = connection.getRepository(GroupCategoryReport);
     this.eventRepository = connection.getRepository(GroupEvent);
+    this.groupCategoryRepository = connection.getRepository(GroupCategory);
   }
 
   async findAll(req: SearchDto): Promise<any[]> {
@@ -79,51 +80,90 @@ export class GroupsService {
     } as any;
   }
 
-  async combo(req: GroupSearchDto): Promise<Group[]> {
-    console.log('GroupSearchDto', req);
-    const findOps: FindConditions<Group> = {};
-    if (hasValue(req.categories)) {
-      findOps.categoryId = In(req.categories);
+  async combo(req: GroupSearchDto, user: any): Promise<Group[]> {
+    const findOps: Record<string, any> = {};
+
+    if (hasValue(user)) {
+      const groupIds =
+        await this.groupsPermissionsService.getUserGroupIds(user);
+      findOps.id = In(groupIds);
     }
+
+    if (hasValue(req.categories)) {
+      const categoryIds = [];
+
+      for (const categoryName of Array.isArray(req.categories)
+        ? req.categories
+        : [req.categories]) {
+        const groupCategory = await this.groupCategoryRepository.findOne({
+          where: { name: categoryName },
+        });
+        categoryIds.push(groupCategory.id);
+      }
+
+      findOps.category = { id: In(categoryIds) };
+    }
+
     if (hasValue(req.query)) {
       findOps.name = ILike(`%${req.query}%`);
     }
-    console.log('findOps', findOps);
-    return await this.treeRepository.find({
-      select: ['id', 'name', 'categoryId', 'parent'],
+
+    const data = await this.treeRepository.find({
+      select: ["id", "name", "category", "parent"],
       where: findOps,
       skip: req.skip,
       take: req.limit,
       cache: true,
     });
+
+    return data;
   }
 
-  async create(data: CreateGroupDto) {
+  async create(
+    data: CreateGroupDto,
+    user: any,
+    seedingDatabase: boolean = false,
+  ) {
     Logger.log(`Create.Group starting ${data.name}`);
     let place: GooglePlaceDto = null;
     if (data.address?.placeId) {
       place = await this.googleService.getPlaceDetails(data.address.placeId);
     }
 
-    const toSave = new Group();
-    toSave.name = data.name;
-    toSave.privacy = data.privacy;
-    toSave.metaData = data.metaData;
-    toSave.categoryId = data.categoryId;
-    toSave.address = place;
-    toSave.details = data.details;
-    toSave.parent = data.parentId
-      ? await this.treeRepository.findOne(data.parentId)
-      : null;
-    const result = await this.treeRepository.save(toSave);
+    if (hasValue(data.parentId) && !seedingDatabase) {
+      await this.groupsPermissionsService.assertPermissionForGroup(
+        user,
+        data.parentId,
+      );
+    }
 
-    return this.findOne(result.id, true);
+    const newGroupCategory = await this.groupCategoryRepository.findOne({
+      where: {
+        name: data.categoryName,
+      },
+    });
+    const newGroup = new Group();
+    newGroup.name = data.name;
+    newGroup.privacy = data.privacy;
+    newGroup.metaData = data.metaData;
+    newGroup.category = newGroupCategory;
+    newGroup.address = place;
+    newGroup.details = data.details;
+    newGroup.parent = data.parentId
+      ? await this.treeRepository.findOne({ where: { id: data.parentId } })
+      : null;
+
+    return await this.treeRepository.save(newGroup);
   }
 
-  async findOne(id: number, full = true) {
-    const data = await this.treeRepository.findOne(id, {
-      relations: ['category', 'parent'],
+  async findOne(id: number, full = true, user: any = null) {
+    const data = await this.treeRepository.findOne({
+      where: { id },
+      relations: ["category", "parent"],
     });
+    if (!data) {
+      return null;
+    }
     Logger.log(`Read.Group success id:${id}`);
     if (full) {
       Logger.log(`Read.Group loading full scope id:${id}`);
@@ -144,7 +184,7 @@ export class GroupsService {
         totalMem = 0;
       await (
         await this.eventRepository.find({
-          relations: ['attendance', 'group', 'group.members'],
+          relations: ["attendance", "group", "group.members"],
           where: filter,
         })
       ).forEach((it) => {
@@ -156,14 +196,15 @@ export class GroupsService {
 
       const membership = await this.membershipRepository.find({
         where: { role: GroupRole.Leader, groupId: id },
-        select: ['contactId'],
+        select: ["contactId"],
       });
       groupData.leaders = membership.map((it) => it.contactId);
-
+      groupData.canEditGroup =
+        await this.groupsPermissionsService.hasPermissionForGroup(user, id);
       groupData.reports = await this.eventRepository.find({
-        relations: ['category', 'attendance'],
+        relations: ["category", "attendance"],
         where: { groupId: In(groupData.children) },
-        select: ['id', 'name', 'startDate'],
+        select: ["id", "name", "startDate"],
       });
 
       return groupData;
@@ -173,17 +214,20 @@ export class GroupsService {
 
   async update(
     dto: UpdateGroupDto,
+    user: any,
   ): Promise<GroupListDto | GroupDetailDto | any> {
     Logger.log(`Update.Group groupID:${dto.id} starting`);
 
+    await this.groupsPermissionsService.assertPermissionForGroup(user, dto.id);
+
     const currGroup = await this.repository
       .createQueryBuilder()
-      .where('id = :id', { id: dto.id })
+      .where("id = :id", { id: dto.id })
       .getOne();
 
     if (!currGroup)
       throw new ClientFriendlyException(`Invalid group ID:${dto.id}`);
-    let place: GooglePlaceDto = null;
+    let place: GooglePlaceDto;
     if (dto.address && dto.address.placeId !== currGroup.address?.placeId) {
       Logger.log(`Update.Group groupID:${dto.id} fetching coordinates`);
       place = await this.googleService.getPlaceDetails(dto.address.placeId);
@@ -194,10 +238,26 @@ export class GroupsService {
 
     let parentGroup = null;
     if (hasValue(dto.parentId)) {
-      parentGroup = await this.treeRepository.findOne(dto.parentId);
+      parentGroup = await this.treeRepository.findOne({
+        where: { id: dto.parentId },
+      });
+    }
+    let category = currGroup.category;
+    if (hasValue(dto.categoryId)) {
+      category = await this.groupCategoryRepository.findOne({
+        where: {
+          id: dto.categoryId,
+        },
+      });
+    }
+    if (hasValue(dto.categoryName)) {
+      category = await this.groupCategoryRepository.findOne({
+        where: {
+          name: dto.categoryName,
+        },
+      });
     }
 
-    console.log(`Update.Group Id:${dto.parentId} parentGroup: `, parentGroup);
     const result = await this.repository
       .createQueryBuilder()
       .update(Group)
@@ -206,10 +266,10 @@ export class GroupsService {
         parent: parentGroup,
         details: dto.details,
         privacy: dto.privacy,
-        categoryId: dto.categoryId,
+        category: category,
         address: place,
       })
-      .where('id = :id', { id: dto.id })
+      .where("id = :id", { id: dto.id })
       .execute();
     if (result.affected)
       Logger.log(
@@ -218,7 +278,8 @@ export class GroupsService {
     return await this.findOne(dto.id, true);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, user: any): Promise<void> {
+    await this.groupsPermissionsService.assertPermissionForGroup(user, id);
     await this.treeRepository.delete(id);
   }
 
@@ -229,5 +290,13 @@ export class GroupsService {
 
   async count(): Promise<number> {
     return await this.repository.count();
+  }
+
+  async getGroupsByCategory(categoryName: string): Promise<Group[]> {
+    return this.repository
+      .createQueryBuilder("group")
+      .leftJoinAndSelect("group.category", "category")
+      .where("category.name = :categoryName", { categoryName })
+      .getMany();
   }
 }
